+++ tests/Constants.sol
@@ -16,6 +16,6 @@ library Constants {
   uint24 public constant MAX_ALLOWED_COLLATERAL_RISK = 1000_00; // 1000.00%
   uint256 public constant MAX_ALLOWED_DYNAMIC_CONFIG_KEY = type(uint24).max;
   uint256 public constant MAX_ALLOWED_ASSET_ID = type(uint16).max;
-  uint24 public constant MAX_USER_COLLATERALS = type(uint24).max;
-  uint24 public constant MAX_USER_BORROWS = type(uint24).max;
+  uint16 public constant MAX_USER_COLLATERALS = type(uint16).max;
+  uint16 public constant MAX_USER_BORROWS = type(uint16).max;
 }
