+++ tests/mocks/MockSpokeInstance.sol
@@ -27,8 +27,8 @@ contract MockSpokeInstance is Spoke {
     __AccessManaged_init(authority);
     if (_spokeConfig.targetHealthFactor == 0) {
       _spokeConfig.targetHealthFactor = HEALTH_FACTOR_LIQUIDATION_THRESHOLD;
-      _spokeConfig.maxUserCollaterals = type(uint24).max;
-      _spokeConfig.maxUserBorrows = type(uint24).max;
+      _spokeConfig.maxUserCollaterals = type(uint16).max;
+      _spokeConfig.maxUserBorrows = type(uint16).max;
       emit UpdateSpokeConfig(_spokeConfig);
     }
   }
